moving CoreseException from next.data to next.util
